requirements.txt
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
Pillow>=10.4.0
numpy>=2.1
pandas>=2.2.3
python-dotenv>=1.0.1
tqdm>=4.66.4
